+++ src/editortextbox.cpp
@@ -315,8 +315,8 @@ void EditorTextBox::draw(NVGcontext* ctx) {
       if (format_string.length()) {
         if (value_type == Value::t_integer) {// integer
           char buf[20];
-          int val = std::atoi(p);
-          snprintf(buf, 20, format_string.c_str(), val / scale);
+          long val = std::atol(p);
+          snprintf(buf, 20, format_string.c_str(), (long)(val / scale));
           valStr = buf;
         }
         else if (value_type == Value::t_float) {
@@ -334,8 +334,8 @@ void EditorTextBox::draw(NVGcontext* ctx) {
       }
       else if (value_type == Value::t_integer) {
         char buf[20];
-        int val = std::atoi(p);
-        snprintf(buf, 20, "%d", (int)(val / scale));
+        long val = std::atol(p);
+        snprintf(buf, 20, "%ld", (long)(val / scale));
         valStr = buf;
       }
     }
